requirements.txt
telethon>=1.34.0
openai>=1.0.0
python-dotenv>=1.0.1
pyyaml>=6.0
redis>=4.2.0
